ref(charts): Make zoom reusable in charts
This allows any chart to be able to zoom
