+++ src/sentry/static/sentry/app/components/charts/utils.jsx
@@ -1,3 +1,5 @@
+import moment from 'moment';
+
 const DEFAULT_TRUNCATE_LENGTH = 80;
 
 export function truncationFormatter(value, truncate) {
@@ -9,4 +11,14 @@ export function truncationFormatter(value, truncate) {
   return value.length > truncationLength
     ? value.substring(0, truncationLength) + '…'
     : value;
+}
+
+export function useHourlyInterval(datetimeObj) {
+  const {period, start, end} = datetimeObj;
+
+  if (typeof period === 'string') {
+    return period.endsWith('h') || period === '1d';
+  }
+
+  return moment(end).diff(start, 'hours') <= 24;
 }
